cookbook/requirements.txt
Django==1.11.13
graphene_django==2.2.0
django-graphql-jwt==0.2.0
django-filter==1.1.0